added types files
